requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.2
APScheduler==3.8.1
async-timeout==4.0.1
asyncio==3.4.3
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==7.1.2
cryptography==36.0.0
decorator==5.1.0
Deprecated==1.2.13
django-environ==0.8.1
environ==1.0
firebase==3.0.1
firebase-admin==5.1.0
Flask==2.0.2
Flask-APScheduler==1.12.2
flask-ngrok==0.0.25
frozenlist==1.2.0
future==0.18.2
google-api-core==2.3.0
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.42.0
grpcio-status==1.42.0
gunicorn==20.1.0
h11==0.12.0
httplib2==0.20.2
httptools==0.3.0
idna==3.3
ipdb==0.13.9
ipython==7.30.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jwcrypto==1.0
line-bot-sdk==2.0.1
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
msgpack==1.0.3
multidict==5.2.0
nested-lookup==0.2.23
numpy==1.21.4
packaging==21.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.23
proto-plus==1.19.8
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-jwt==3.3.1
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
rsa==4.8
schedule==1.1.0
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
traitlets==5.1.1
typing-extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.13.4
uvloop==0.16.0
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
yarl==1.7.2
